Move spin functions back to lib.rs
